backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
psycopg2-binary
alembic
httpx
pydantic-settings
